refactor(libs): Move methods out of ContentfulService
